requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.8.0
lightning==2.5.4
torchvision==0.23.0
transformers==4.51.3
hydra-core==1.3.2
modelscope==1.29.2
timm==0.4.12
flash-attn==2.8.3
nvidia-ml-py
ninja
viser
kornia
cupy-cuda12x
opencv-python
tensorboard
imageio
imageio[ffmpeg]
imageio[pyav]
safetensors
einops
sentencepiece
protobuf
ftfy
omegaconf
websockets
pandas
wandb
click
numpy
tqdm
rerun-sdk
matplotlib
python-pycg
natsort
gdown
OpenEXR<3.3.0
ray